requirements.txt
Cartopy==0.18.0
gmsh==4.10.3
h5py==2.10.0
ipod.egg==info
joblib==1.0.0
Keras==2.4.3
matplotlib==3.5.1
netCDF4==1.5.8
numpy==1.20.3
openquake.engine==3.11.3
pandas==1.0.5
paramiko==2.8.1
pod==0.54.2
pyproj==2.6.1.post1
PyYAML==6.0
rasterio==1.1.2
scikit_learn==1.1.1
scipy==1.5.4
seissolxdmf==0.1.0
seissolxdmfwriter==0.1.0
setuptools==58.0.4
shakemap.egg==info
tqdm==4.56.0
trimesh==3.10.7
